refactor exceptions to a single file
