+++ include/exceptions.hpp
@@ -28,3 +28,17 @@ public:
 private:
     std::string message_;
 };
+
+class InvalidDataTypeException : public std::exception
+{
+public:
+    InvalidDataTypeException(const std::string& message) : message_(message) {}
+
+    const char* what() const noexcept override
+    {
+        return message_.c_str();
+    }
+
+private:
+    std::string message_;
+};
